+++ redvypr/CHANGELOG.md
@@ -13,8 +13,9 @@ redvypr changelog
 - replaced `pkg_resources` with a modern alternative
 - bug and layout in `redvyprSubscribeWidget`
 - `RedvyprMultipleAddressesWidget` with new functionality of `self.address_names`, a dictionary of names that are assigned with `RedvyprAddress`
-- `new_data(datapacket)` signal properly implemented in `redvypr.device`
-  - first usage in device `devices.event.event.EventConfigEditor` 
+- `new_data([datapacket])` signal properly implemented in `redvypr.device`
+  - first usage in device `devices.event.event.EventConfigEditor`
+  - Plan is to replace the `gui_widget.update_data(data)` functions with signal functionality
 
 ---